+++ dev/preview.html
@@ -25,6 +25,6 @@
             title: 'Acme Storefront API',
         };
     </script>
-    <script src="../resources/ui/app.js" defer></script>
+    <script type="module" src="../resources/ui/app.js"></script>
 </body>
 </html>
