+++ resources/ui/app.css
@@ -353,11 +353,12 @@ a { color: inherit; }
     align-items: flex-start;
     gap: 10px;
     width: 100%;
+    height: 100%;
     text-align: left;
     background: none;
     border: 0;
     border-radius: 8px;
-    padding: 8px 10px;
+    padding: 9px 10px;
     cursor: pointer;
     color: var(--text-2);
     overflow: hidden;
@@ -392,13 +393,23 @@ a { color: inherit; }
     text-transform: uppercase;
     color: var(--blue);
 }
-.nav-item__main { min-width: 0; display: grid; gap: 3px; }
-.nav-item__path { font-family: var(--mono); font-size: 12.5px; line-height: 1.35; word-break: break-all; }
+.nav-item__main { min-width: 0; display: grid; align-content: start; gap: 3px; }
+.nav-item__path {
+    display: -webkit-box;
+    overflow: hidden;
+    font-family: var(--mono);
+    font-size: 12.5px;
+    line-height: 1.35;
+    overflow-wrap: anywhere;
+    word-break: normal;
+    -webkit-box-orient: vertical;
+    -webkit-line-clamp: 2;
+}
 .nav-item__summary {
     display: block;
     color: var(--muted);
     font-size: 12px;
-    line-height: 1.3;
+    line-height: 1.25;
     overflow: hidden;
     text-overflow: ellipsis;
     white-space: nowrap;
